criacao de arquivos bubblesor para float e double crescente

+++ Bubblesort/c/bubbleDoubleCrescente.c
@@ -7,7 +7,7 @@ Parametros: um vetor de inteiros (vetor que tera os
 elementos trocados) e dois inteiros (indice dos 
 dois elementos a serem trocados)
 */
-void swap (float vet[], int i, int j) {
+void swap (int vet[], int i, int j) {
     int aux = vet[i];
     vet[i] = vet[j];
     vet[j] = aux;
@@ -23,7 +23,7 @@ void bubblesort (int vet[]) {
     //Calculo do tamanho do vetor
     int n = sizeof(vet)/sizeof(int);
 
-    //Ordenaacao do vetor
+    //Ordenacao do vetor
     for (int i = n-1; i > 0; i--) {
         for (int j = 0; j < i; j++) {
             if (vet[j] > vet[j+1]) {
